refactor: split cache application-path helpers into extension
Merging main's dev-binary ACL hardening into this branch pushed the
KeychainCacheStore enum body past SwiftLint's 800-line limit; the
application-path/bundle-resolution helpers move to a dedicated file.
